+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/AICommitAction.kt
@@ -1,8 +1,10 @@
 package com.github.blarc.ai.commits.intellij.plugin
 
+import com.github.blarc.ai.commits.intellij.plugin.AICommitsBundle.message
 import com.github.blarc.ai.commits.intellij.plugin.notifications.Notification
 import com.github.blarc.ai.commits.intellij.plugin.notifications.sendNotification
 import com.github.blarc.ai.commits.intellij.plugin.settings.ProjectSettings
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
@@ -12,12 +14,35 @@ import com.intellij.openapi.vcs.ui.CommitMessage
 import com.intellij.vcs.commit.AbstractCommitWorkflowHandler
 
 class AICommitAction : AnAction(), DumbAware {
+
+
+    override fun getActionUpdateThread(): ActionUpdateThread {
+        return ActionUpdateThread.EDT
+    }
+
+    override fun update(e: AnActionEvent) {
+        e.project?.service<ProjectSettings>()?.getActiveLLMClientConfiguration()?.let {
+            if (it.getGenerateCommitMessageJob()?.isActive == true) {
+                e.presentation.icon = Icons.Process.STOP.getThemeBasedIcon()
+            } else {
+                e.presentation.icon = it.getClientIcon()
+                e.presentation.text = message("action.tooltip", it.name)
+            }
+        }
+    }
+
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return
 
         val llmClient = project.service<ProjectSettings>().getActiveLLMClientConfiguration()
         if (llmClient == null) {
             sendNotification(Notification.clientNotSet())
+            return
+        }
+
+        val generateCommitMessageJob = llmClient.getGenerateCommitMessageJob()
+        if (generateCommitMessageJob?.isActive == true) {
+            generateCommitMessageJob.cancel()
             return
         }