+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/Icons.kt
@@ -27,4 +27,8 @@ object Icons {
     val AZURE_OPEN_AI = AICommitsIcon("/icons/azureOpenAi.svg", null)
     val HUGGING_FACE = AICommitsIcon("/icons/huggingface.svg", null)
     val GITHUB = AICommitsIcon("/icons/github15bright.svg", "/icons/github15dark.svg")
+
+    object Process {
+        val STOP = AICommitsIcon("/icons/stop.svg", "/icons/stop_dark.svg")
+    }
 }