+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/clients/azureOpenAi/AzureOpenAiClientConfiguration.kt
@@ -8,6 +8,7 @@ import com.intellij.openapi.vcs.ui.CommitMessage
 import com.intellij.util.xmlb.annotations.Attribute
 import com.intellij.util.xmlb.annotations.Transient
 import com.intellij.vcs.commit.AbstractCommitWorkflowHandler
+import kotlinx.coroutines.Job
 import javax.swing.Icon
 
 class AzureOpenAiClientConfiguration : LLMClientConfiguration(
@@ -47,8 +48,8 @@ class AzureOpenAiClientConfiguration : LLMClientConfiguration(
         return AzureOpenAiClientService.getInstance().generateCommitMessage(this, commitWorkflowHandler, commitMessage, project)
     }
 
-    override fun cancelGenerateCommitMessage() {
-        AzureOpenAiClientService.getInstance().cancelGenerateCommitMessage()
+    override fun getGenerateCommitMessageJob(): Job? {
+        return AzureOpenAiClientService.getInstance().generateCommitMessageJob
     }
 
     // Model names are retrieved from Enum and do not need to be refreshed.