+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/clients/qianfan/QianfanClientConfiguration.kt
@@ -8,6 +8,7 @@ import com.intellij.util.xmlb.annotations.Attribute
 import com.intellij.util.xmlb.annotations.Transient
 import com.intellij.vcs.commit.AbstractCommitWorkflowHandler
 import dev.langchain4j.model.qianfan.QianfanChatModelNameEnum
+import kotlinx.coroutines.Job
 import javax.swing.Icon
 
 class QianfanClientConfiguration : LLMClientConfiguration(
@@ -48,8 +49,8 @@ class QianfanClientConfiguration : LLMClientConfiguration(
         return QianfanClientService.getInstance().generateCommitMessage(this, commitWorkflowHandler, commitMessage, project)
     }
 
-    override fun cancelGenerateCommitMessage() {
-        QianfanClientService.getInstance().cancelGenerateCommitMessage()
+    override fun getGenerateCommitMessageJob(): Job? {
+        return QianfanClientService.getInstance().generateCommitMessageJob
     }
 
     // Model names are retrieved from Enum and do not need to be refreshed.