+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - More options for configuring LLM clients.
+- **Use the chosen LLM client icon as the generate commit message action's icon.**
+- Option to stop the commit message generation by clicking the action icon again.
 
 ### Fixed